backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
supabase==2.9.1
twilio==9.3.0
pyjwt==2.9.0
httpx==0.27.0
python-multipart==0.0.12
